add: remove "add.interactive.useBuiltin" & Perl "git add--interactive"
Since [1] first released with Git v2.37.0 the built-in version of "add
-i" has been the default. That built-in implementation was added in
[2], first released with Git v2.25.0.

At this point enough time has passed to allow for finding any
remaining bugs in this new implementation, so let's remove the
fallback code.

As with similar migrations for "stash"[3] and "rebase"[4] we're
keeping a mention of "add.interactive.useBuiltin" in the
documentation, but adding a warning() to notify any outstanding users
that the built-in is now the default. As with [5] and [6] we should
follow-up in the future and eventually remove that warning.

1. 8b81a78473a (add -i: default to the built-in implementation,
   2021-11-30)
2. c594f2d3f8b (Start to implement a built-in version of `git add
   --interactive`, 2019-11-13)
3. e5b0dba8c32 (stash: remove the stash.useBuiltin setting,
   2020-03-03)
4. b8b9bb2694b (rebase: remove the rebase.useBuiltin setting,
   2019-03-18)
5. e3723fc92dc (stash: remove documentation for `stash.useBuiltin`,
   2022-01-27)
6. 87b6fe31476 (rebase: remove transitory rebase.useBuiltin setting &
   env, 2021-03-23)

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>